refactor(util,typescript-folder): created new folder `util`, moved typescript files to config
