+++ src/base/config/eslint.ts
@@ -1,8 +1,8 @@
 import { TextFile } from 'projen';
 import { Config } from '../config';
-import { ProjectTypes } from '../../types/project';
-import { isValidProject } from '../../utils';
-import { ConfigFile } from '../../types';
+import { ProjectTypes } from '../../util/types/project';
+import { isValidProject } from '../../util/utils';
+import { ConfigFile } from '../../util/types';
 
 /**
  * Base class for implementing all relevant EsLint configuration.
