+++ src/base/config/github.ts
@@ -1,7 +1,7 @@
 import { TextFile } from 'projen';
 import { Config } from '../config';
-import { ProjectTypes } from '../../types/project';
-import { isValidProject } from '../../utils';
+import { ProjectTypes } from '../../util/types/project';
+import { isValidProject } from '../../util/utils';
 
 /**
  * Base class for implementing all relevant GitHub configuration.
