+++ src/base/config/prettier.ts
@@ -1,8 +1,8 @@
 import { JsonFile, TextFile } from 'projen';
 import { Config } from '../config';
 import { TrailingComma } from 'projen/lib/javascript';
-import { ConfigFile, ProjectTypes } from '../../types';
-import { isValidProject } from '../../utils';
+import { ConfigFile, ProjectTypes } from '../../util/types';
+import { isValidProject } from '../../util/utils';
 
 /**
  * Base class for implementing all relevant Prettier configuration.
