+++ src/cap-service/config/samplecode.ts
@@ -1,7 +1,7 @@
 import { SampleFile } from 'projen';
 import { CapServiceProject, CapServiceProjectOptions } from '.././project';
 import { BaseProject, SampleCodeConfigBase } from '../../base';
-import { Settings } from '../../types';
+import { Settings } from '../../util/types';
 
 /**
  * Implementing all relevant SampleCode configuration for the CapService project.
