+++ src/github-action/config/commitlint.ts
@@ -1,5 +1,5 @@
 import { CommitLintConfigBase } from '../../base/config/commitlint';
-import { Settings } from '../../types';
+import { Settings } from '../../util/types';
 
 /**
  * Implementing all relevant CommitLint configuration for the GitHubAction project.
