+++ src/github-action/config/npm.ts
@@ -1,6 +1,6 @@
 import { NpmConfigBase } from '../../base/config/npm';
 import { BaseProject } from '../../base/project';
-import { Settings } from '../../types';
+import { Settings } from '../../util/types';
 
 /**
  * Implementing all relevant NPM configuration for the GitHubAction project.
