+++ src/github-action/index.ts
@@ -1,5 +1,5 @@
 export * from './config/commitlint';
 export * from './config/npm';
 export * from './project';
-export * from './typescript';
+export * from './config/typescript';
 export * from './config/samplecode';
