+++ src/index.ts
@@ -2,5 +2,5 @@ export * from './base';
 export * from './jsii';
 export * from './cap-service';
 export * from './github-action';
-export * from './types';
-export * from './utils';
+export * from './util/types';
+export * from './util/utils';
