+++ src/jsii/config/npm.ts
@@ -1,5 +1,5 @@
 import { NpmConfigBase } from '../../base/config/npm';
-import { Settings } from '../../types';
+import { Settings } from '../../util/types';
 import { JsiiProject } from '.././project';
 
 /**
