+++ src/jsii/index.ts
@@ -1,3 +1,3 @@
 export * from './config/npm';
 export * from './project';
-export * from './typescript';
+export * from './config/typescript';
